Create bikeshare.sql
This file creates a database for the bikeshare data. The total_share will merge them all together while keeping the quarter and year as a variable. Something to note here is that for Q1 and Q2 of 2012 they added a duration in seconds field, then quickly got rid of the field.